--- sampleProject.swift/HistoryListViewController.swift
@@ -1,30 +1,0 @@
-//
-//  HistoryListViewController.swift
-//  sampleProject
-//
-//  Created by TeruakiEnomoto on 23/06/2019.
-//  Copyright © 2019 TeruakiEnomoto. All rights reserved.
-//
-
-import UIKit
-
-class HistoryListViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
