+++ sampleProject.swift/RecordViewController.swift
@@ -18,15 +18,15 @@ class RecordViewController: UIViewController {
     @IBOutlet weak var positiveTextField3: UITextField!
     @IBOutlet weak var positiveTextField4: UITextField!
     
-//    対象となるnegativeをnilで定義
+    //    対象となるnegativeをnilで定義
     var negative: Negative? = nil
-//    対象となるpositiveをnilで定義
+    //    対象となるpositiveをnilで定義
     var positive: Positive? = nil
     
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
-//        .realmファイルが保存されている場所を表示する
+        //        .realmファイルが保存されている場所を表示する
         print(Realm.Configuration.defaultConfiguration.fileURL!)
     }
     
@@ -36,16 +36,16 @@ class RecordViewController: UIViewController {
         let positiveEvent2 = positiveTextField2.text
         let positiveEvent3 = positiveTextField3.text
         let positiveEvent4 = positiveTextField4.text
-//        negativeが入力されていてpositiveが空欄の場合
+        //        negativeが入力されていてpositiveが空欄の場合
         if negativeTextField.text != "" && positiveTextField1.text == "" && positiveTextField2.text == "" && positiveTextField3.text == "" && positiveTextField4.text == "" {
             let alert = UIAlertController(title: "入力不足", message: "ポジティブ欄が入力されていません。入力してください。", preferredStyle: .alert)
-//            選択肢
+            //            選択肢
             let yesAction = UIAlertAction(title: "OK", style: .default) {
                 (UIAlertAction) in print("OK")
             }
-//                選択肢の追加
+            //                選択肢の追加
             alert.addAction(yesAction)
-//                アラートを表示する
+            //                アラートを表示する
             present(alert, animated: true, completion: nil)
         } else if negativeTextField.text == "" && positiveTextField1.text == "" && positiveTextField2.text == ""  && positiveTextField3.text == "" && positiveTextField4.text == "" {   //      negativeとpositiveの両方が空欄の場合
             let alert = UIAlertController(title: "入力不足", message: "ネガティブ欄とポジティブ欄が入力されていません。入力してください。", preferredStyle: .alert)
@@ -59,9 +59,9 @@ class RecordViewController: UIViewController {
             present(alert, animated: true, completion: nil)
             
         } else {    //  Realmにデータを保存する
-//            recordNegativeメソッドを呼び出す
+            //            recordNegativeメソッドを呼び出す
             recordNegative(with: negativeEvent!)
-//            recordPositiveメソッドを呼び出す
+            //            recordPositiveメソッドを呼び出す
             recordPositive(with: positiveEvent1!)
             if positiveTextField2.text != "" {
                 recordPositive(with: positiveEvent2!)
@@ -72,7 +72,7 @@ class RecordViewController: UIViewController {
             if positiveTextField4.text != "" {
                 recordPositive(with: positiveEvent4!)
             }
-//            入力完了のアラート
+            //            入力完了のアラート
             let completeAlert = UIAlertController(title: "入力完了", message: "おめでとうございます。記録できましたね!!", preferredStyle: .alert)
             //            選択肢
             let completeAction = UIAlertAction(title: "OK", style: .default) {
@@ -84,60 +84,60 @@ class RecordViewController: UIViewController {
             present(completeAlert, animated: true, completion: nil)
         }
         
-//        テキストフィールドの値を初期化
+        //        テキストフィールドの値を初期化
         negativeTextField.text = ""
         positiveTextField1.text = ""
         positiveTextField2.text = ""
         positiveTextField3.text = ""
         positiveTextField4.text = ""
-
+        
     }
-
+    
 }
 
 //  Realmに関する処理
 extension RecordViewController {
-//    NegativeクラスはNegative.swiftで定義済み
-//    Negative作成
+    //    NegativeクラスはNegative.swiftで定義済み
+    //    Negative作成
     func recordNegative(with text: String) {
-//        Realmに接続
+        //        Realmに接続
         let realm = try! Realm()
-//        登録処理
-//        Negativeの初期化
+        //        登録処理
+        //        Negativeの初期化
         negative = Negative()
         negative?.id = getNegativeMaxId()
         negative?.negativeText = text
         negative?.date = Date()
         
-//        Realmに書き込み
+        //        Realmに書き込み
         try! realm.write {      //  Realmに書き込みモード
             realm.add(negative!)        //  DBに追加
         }
     }
-//        getNegativeMaxId
+    //        getNegativeMaxId
     func getNegativeMaxId() -> Int {
         let realm = try! Realm()
         return (realm.objects(Negative.self).max(ofProperty:"id")as Int? ?? 0) + 1
     }
     
     
-//    Positive作成
-//    PositiveクラスはPositive.swiftで定義済み
+    //    Positive作成
+    //    PositiveクラスはPositive.swiftで定義済み
     func recordPositive(with text: String){
-//        Realmに接続
+        //        Realmに接続
         let realm = try! Realm()
-//        登録処理
-//        Positiveの初期化
+        //        登録処理
+        //        Positiveの初期化
         positive = Positive()
         positive?.id = getPositiveMaxId()
         positive?.positiveText = text
         positive?.date = Date()
-//        Realmに書き込み
+        //        Realmに書き込み
         try! realm.write{       //  Realmの書き込みモード
             realm.add(positive!)        //  DBに追加
         }
     }
-//    getPositiveMaxId
+    //    getPositiveMaxId
     func getPositiveMaxId() -> Int {
         let realm = try! Realm()
         return (realm.objects(Positive.self).max(ofProperty: "id") as Int? ?? 0) + 1
