+++ sampleProject.swift/RegistViewController.swift
@@ -21,24 +21,24 @@ class RegistViewController: UIViewController {
         super.viewDidLoad()
         
     }
-
+    
     @IBAction func didClickBtn(_ sender: Any) {
         let userName = userNameTextField.text
-//        テキストフィールドの値が空じゃない場合
+        //        テキストフィールドの値が空じゃない場合
         if userNameTextField.text != "" {
-//            registUserメソッドを呼び出す
+            //            registUserメソッドを呼び出す
             registUser(with: userName!)
-//            画面遷移
+            //            画面遷移
             performSegue(withIdentifier: "toStart", sender: userName)
         } else { // 値が空の場合はアラートを表示する
             let alert = UIAlertController(title: "未入力", message: "ユーザー名が入力されていません", preferredStyle: .alert)
-//            選択肢
+            //            選択肢
             let yesAction = UIAlertAction(title: "OK", style: .default) {
                 (UIAlertAction) in print("OK")
             }
-//            選択肢の追加
+            //            選択肢の追加
             alert.addAction(yesAction)
-//            アラートを表示する
+            //            アラートを表示する
             present(alert, animated: true, completion: nil)
             
         }
@@ -47,28 +47,28 @@ class RegistViewController: UIViewController {
 
 //  Realmに関する処理
 extension RegistViewController {
-//    UserクラスはUser.swiftで定義済み
-//    User作成
+    //    UserクラスはUser.swiftで定義済み
+    //    User作成
     func registUser(with name:String) {
-//        Realmに接続
+        //        Realmに接続
         let realm = try! Realm()
-//        .realmファイルが保存されている場所を表示する
+        //        .realmファイルが保存されている場所を表示する
         print(Realm.Configuration.defaultConfiguration.fileURL!)
         
-//        登録処理
-//        USERの初期化
+        //        登録処理
+        //        USERの初期化
         user = User()
         user?.id = getMaxId()
         user?.name = name
         user?.date = Date()
         
-//        Realmに書き込み
+        //        Realmに書き込み
         try! realm.write{   // Realmの書き込みモード
             realm.add(user!)    // DBに追加
         }
     }
     
-//    getMaxId
+    //    getMaxId
     func getMaxId() -> Int {
         let realm = try! Realm()
         return (realm.objects(User.self).max(ofProperty: "id") as Int? ?? 0) + 1
